refactor(LinkPage): Move into /pages/sharing
